requirements.txt
streamlit>=1.24.0
streamlit-webrtc>=0.64.5
opencv-python-headless
pillow
numpy
# Optional for server-side Keras inference:
tensorflow>=2.12.0